app/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
transformers==4.41.1
torch==2.3.0
pillow==10.3.0
requests==2.32.2
